requirements.txt
aliyun-python-sdk-core-v3==2.13.33
amqp==5.1.1
anyio==3.6.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==20.3.0
autobahn==22.6.1
Automat==20.2.0
Babel==2.10.3
billiard==3.6.4.0
celery==5.2.7
certifi==2020.12.5
cffi==1.15.1
channels==3.0.5
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.22.5
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.15
django-author==1.2.0
django-cors-headers==3.13.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.5.0
django-filter==22.1
django-formtools==2.3
django-guardian==2.3.0
django-import-export==2.5.0
django-oauth-toolkit==2.1.0
django-redis==5.2.0
django-reversion==3.0.9
django-simpleui==2022.6.131
djangorestframework==3.12.4
djangorestframework-camel-case==1.3.0
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
dlib==19.24.0
dnspython==2.2.1
docutils==0.19
drf-excel==2.1.0
drf-extensions==0.7.1
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.8.1
drf-yasg @ file:///E:/2022%E5%B9%B4%E4%BB%A3%E7%A0%81%E6%B1%87%E6%80%BB/drf-yasg
et-xmlfile==1.1.0
eventlet==0.33.1
future==0.18.2
ghp-import==2.1.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httpie==2.4.0
httplib2==0.19.1
httptools==0.4.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.12.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
jwcrypto==1.3.1
kombu==5.2.4
livereload==2.6.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.1.1
mergedeep==1.3.4
mixins==0.1.4
mkdocs==1.3.1
mkdocs-material==8.3.9
mkdocs-material-extensions==1.0.3
mysqlclient==2.0.3
numpy==1.23.1
oauthlib==3.2.0
odfpy==1.4.1
openapi==1.1.0
opencv-python==4.6.0.66
openpyxl==3.0.7
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pymdown-extensions==9.5
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-gettext==4.0
pytz==2022.2.1
PyYAML==5.4.1
pyyaml_env_tag==0.1
redis==4.3.4
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
serializers==0.2.4
service-identity==21.1.0
setuptools-git==1.2
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
swagger-spec-validator==2.7.4
tablib==3.0.0
tornado==6.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
ua-parser==0.15.0
uritemplate==3.0.1
urllib3==1.26.4
user-agents==2.2.0
uvicorn==0.18.2
vine==5.0.0
watchdog==2.1.9
watchfiles==0.16.1
wcwidth==0.2.5
websockets==10.3
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
zope.interface==5.4.0
